moved users.tsx, deleted engilsh.ts, deleted sidebar
