+++ src/language/en.tsx
@@ -2,5 +2,14 @@ export const lang = {
     login: {
         "email-validation": "Please enter your email address as your login!",
         "password-validation": "Please enter your password!"
+    },
+    username: {
+        "firstName-validation": "Please enter your first name!",
+        "lastName-validation": "Please enter your last name!"
+    },
+    profile:{
+        "profile":"Profile",
+        "logout":"Logout"
     }
+
 }
