--- src/language/english.tsx
@@ -1,10 +1,0 @@
-export const lang = {
-    username: {
-        "firstName-validation": "Please enter your first name!",
-        "lastName-validation": "Please enter your last name!"
-    },
-    profile:{
-      "profile":"Profile",
-      "logout":"Logout"
-    }
-}
