+++ src/views/user/layout.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
-import Sidebar from '~/Components/Sidebar';
+
+
 
 // @ts-ignore
 const Layout = ({children}):JSX.Element => {
     return (
         <div>
-            <Sidebar/><div>{children}</div>
+            <div>{children}</div>
         </div>
     );
 };
